+++ CHANGELOG.md
@@ -1,3 +1,87 @@
+## 1.0.49
+- intl versão 0.20.2
+## 1.0.48
+- nova função isValidJson
+## 1.0.47
+- Ajuste na .toDouble2
+## 1.0.46
+- nova Função arredondarComPrecisao
+## 1.0.45
+- nova Função arredondarComPrecisao
+## 1.0.44
+- nova função toDecimal2UsTruncate
+## 1.0.43
+- upgrade dart sdk <4.0.0
+## 1.0.42
+- fix toCapitalizeTitle bugs.
+## 1.0.41
+- fix toCapitalizeTitle bugs.
+## 1.0.40
+- fix toCapitalizeText para não quebrar linha quando tem ponto final.
+## 1.0.39
+- fix toCapitalizeText para não quebrar linha quando tem ponto final.
+## 1.0.38
+- Novo .toCapitalizeTitle
+## 1.0.37
+- novo .toDouble2 e .toDouble3 para corrigir flutuação de casas decimais
+## 1.0.36
+- ajustes toFusoUtc();
+## 1.0.35
+- nova extension .toCapitalizeString
+- ajuste da função toFusoUtc();
+## 1.0.34
+- nova extension datetime .toMesAnoDescrito -> AGOSTO/2024
+## 1.0.33
+- fix extension validar string de data
+## 1.0.32
+- New extension para validar uma string se é uma data
+## 1.0.31
+- new num? extension, e nova função isEmptyOrNull
+## 1.0.30
+- new String? extension, e nova função isEmptyOrNull
+## 1.0.29
+- fix String extension parseToMap
+## 1.0.28
+- add String extension parseToMap()
+## 1.0.27
+- upgrade package intl para 0.19.0
+## 1.0.26
+- add extension .toLongDate -> 07 de março de 2024 | String
+## 1.0.25
+- Corrigido função .toUtc para não adicionar hora em uma data que já é UTC
+## 1.0.24
+- downgrade package intl para 0.18.1
+## 1.0.23
+- upgrade package intl para 0.19.0
+## 1.0.22
+- upgrade sdk version para 3.5.0
+## 1.0.21
+- nova extension .toUtcFuso(int).
+## 1.0.20
+- fix .toDate String sem hora
+- add String Extension .toFone
+- add String Extension .toCpfCnpj   
+## 1.0.19
+- correção extension DoubleExtension toDecimal2Us para números maior que 999.99
+
+## 1.0.18
+- Dart version 3.3.0
+
+## 1.0.17
+- Converter String para blob;
+
+## 1.0.16
+- Nova extension Blob do Mysql;
+
+## 1.0.15
+- Nova extension double .toDecimalBr(decimalDigits);
+
+## 1.0.14
+- Extensions de double
+
+## 1.0.13
+- nova DateTime extension addMonth
+
 ## 1.0.12
 - nova string extension toCapitalize 
 - nova string extension toCamelCase